services/tasks-service/requirements.txt
flask==3.0.3
psycopg[binary]==3.2.1
kafka-python==2.0.2